+++ internal/lsp/server.go
@@ -3,6 +3,7 @@ package lsp
 import (
 	"context"
 
+	"github.com/lets-cli/lets/internal/env"
 	"github.com/tliron/commonlog"
 	_ "github.com/tliron/commonlog/simple"
 	lsp "github.com/tliron/glsp/protocol_3_16"
@@ -17,6 +18,8 @@ type lspServer struct {
 	version string
 	server  *server.Server
 	storage *storage
+	parser  *parser
+	index   *index
 	log     commonlog.Logger
 }
 
@@ -24,8 +27,22 @@ func (s *lspServer) Run() error {
 	return s.server.RunStdio()
 }
 
+func lspLogVerbosity() int {
+	verbosity := 1
+
+	defer func() {
+		_ = recover()
+	}()
+
+	if env.DebugLevel() > 0 {
+		verbosity = 2
+	}
+
+	return verbosity
+}
+
 func Run(ctx context.Context, version string) error {
-	commonlog.Configure(1, nil)
+	commonlog.Configure(lspLogVerbosity(), nil)
 
 	logger := commonlog.GetLogger(lsName)
 	logger.Infof("Lets LSP server starting %s", version)
@@ -39,6 +56,8 @@ func Run(ctx context.Context, version string) error {
 		version: version,
 		server:  glspServer,
 		storage: newStorage(),
+		parser:  newParser(logger),
+		index:   newIndex(logger),
 		log:     logger,
 	}
 
@@ -50,6 +69,7 @@ func Run(ctx context.Context, version string) error {
 	handler.TextDocumentDidChange = lspServer.textDocumentDidChange
 	handler.TextDocumentDefinition = lspServer.textDocumentDefinition
 	handler.TextDocumentCompletion = lspServer.textDocumentCompletion
+	// TODO: add onDelete
 
 	return lspServer.Run()
 }