+++ internal/lsp/storage.go
@@ -1,6 +1,9 @@
 package lsp
 
+import "sync"
+
 type storage struct {
+	mu        sync.RWMutex
 	documents map[string]*string
 }
 
@@ -11,9 +14,15 @@ func newStorage() *storage {
 }
 
 func (s *storage) GetDocument(uri string) *string {
+	s.mu.RLock()
+	defer s.mu.RUnlock()
+
 	return s.documents[uri]
 }
 
 func (s *storage) AddDocument(uri string, text string) {
+	s.mu.Lock()
+	defer s.mu.Unlock()
+
 	s.documents[uri] = &text
 }